totem/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mysql-connector-python==8.0.16
Pillow==6.0.0
protobuf==3.7.1
pypng==0.0.19
PyQRCode==1.2.1
qrcode==6.1
six==1.12.0
Werkzeug==0.14.1
